feat: Gistの埋め込み要素に<SendWindowSize />を実装した
