+++ packages/zenn-embed-components/src/components/SendWindowSize/index.tsx
@@ -2,6 +2,7 @@ import { useRef } from '@storybook/addons';
 import { PropsWithChildren, useEffect } from 'react';
 
 interface SendWindowSizeProps extends PropsWithChildren<{}> {
+  url: string;
   className?: string;
 }
 
@@ -9,18 +10,19 @@ interface SendWindowSizeProps extends PropsWithChildren<{}> {
  * 他のwindowにウィンドウサイズをpostMessageする
  */
 export const SendWindowSize = ({
+  url,
   children,
   className,
 }: SendWindowSizeProps) => {
   const ref = useRef<HTMLDivElement | null>(null);
 
   useEffect(() => {
+    if (!url) return;
     if (!ref.current) return;
-    if (!window.parent) return;
+    if (window.parent === window) return;
 
     const observer = new ResizeObserver((entries) => {
       entries.forEach((entry) => {
-        const url = window.origin;
         const { width, height } = entry.contentRect;
 
         const data = JSON.stringify({ url, width, height });
@@ -33,7 +35,7 @@ export const SendWindowSize = ({
     observer.observe(ref.current);
 
     return () => observer.disconnect();
-  }, []);
+  }, [url]);
 
   return (
     <div ref={ref} className={className}>
