+++ packages/zenn-embed-components/src/embed/gist/types.ts
@@ -1,0 +1,10 @@
+export type GistApiResponse = {
+  div: string;
+  stylesheet: string;
+};
+
+export interface EmbedGistProps {
+  url?: string;
+  data?: GistApiResponse;
+  error?: Error;
+}
